Addes payment store

+++ database/migrations/2020_10_22_130225_create_payments_table.php
@@ -18,7 +18,7 @@ class CreatePaymentsTable extends Migration
             $table->unsignedBigInteger('booking_id');
             $table->unsignedBigInteger('user_id');
             $table->string('refno');
-            $table->decimal('amount',5, 2);
+            $table->double('amount',10, 2);
             $table->string('currency');
             $table->unsignedBigInteger("uppTransactionId");
             $table->string('pmethod');
